Use RESOLVE_IN_ROOT for path resolving (#222)
There was a regression with our new symlink resolving. Paths that had
components that are symlinks would screw up everything. This happened to
work before because we would hack the symlink string to be inside the
rootfs, but this is obviously a quite bad solution.

Now we use openat2 + RESOLVE_IN_ROOT to resolve file paths relative to
g_rootfs_fd.

We *could* write our own resolving algorithm according to
<https://man7.org/linux/man-pages/man7/path_resolution.7.html>. But this
would:
- need more syscalls
- be easy to screw up 

---

A nice benefit of this solution is it prevents going outside the rootfs
with ie. stat("../../../") or w/e. And solves our hack to make / + .. ==
/

(That being said we are still not a security application so don't get
the wrong idea)