Implement trusted directories (#345)
felix86 could run executables that are inside the rootfs, but not
outside. This is because the entire x86 environment is sandboxed inside
a single directory (the "rootfs"), which is where all x86 libraries and
binaries reside.

However, this can be unfriendly to users who just want to run an x86
executable from wherever it is.

This PR allows running x86 executables from outside the rootfs, as long
as one of their parent directories is a "trusted directory".

---

## Technical details

It's not enough to simply hardlink the executable over to the rootfs
when it's ran: The directory might have important libraries and files
We don't want to mount the trusted directory inside the root: Pollutes
the mounts and requires root. We can't use rootless namespaces for
reasons (messes up with setuid executables like sudo)
We can't symlink: From felix86's perspective, you can't resolve a
symlink to be outside the rootfs

Thus, the decision is that it should be done with a fake mount behavior.
The trusted directory is "mounted" (at least from the perspective of the
emulated executable, **not actually mounted**) in a tmpfs directory.
Whenever felix86 accesses that directory, it's as if it is accessing the
original trusted directory and all its subdirectories and files.

To be exact, the directory is mounted to:
`/run/felix86/trusted/normalized-path/dirname`
where `normalized-path` is the absolute path to the directory with
slashes replaced with hyphens. This is done to distinguish from multiple
directories with the same name. Inside the normalized-path dir exists
the actual `dirname` mounted dir. This is done in case an executable
relies on its parent directory having a certain name.

So for example:
`/home/offtkp/mydir`
would be "mounted" to
`/run/felix86/trusted/-home-offtkp-mydir/mydir`

---

When a file is executed from an untrusted directory, you get a prompt:
<img width="791" height="213" alt="terminalPrompt"
src="https://github.com/user-attachments/assets/5b77e9a0-c933-47f0-8213-dcfc058619b3"
/>
If the program is not ran from a terminal, a GUI dialog is shown:
<img width="672" height="351" alt="guiPrompt"
src="https://github.com/user-attachments/assets/212dff6f-ae3e-4ae3-9982-46f7267cc416"
/>
Once a directory is trusted, further runs won't require authorization.
You can remove the directory from `~/.config/felix86/trusted.txt` to
untrust it.