No longer rely on higher privileges (#364)
Up until now, felix86 will use a setuid application called
felix86-mounter which will handle the parts of felix86 that require root
permission. These included:
- Mounting /dev, /proc, /sys, /run, /tmp
- Bind mounting the rootfs on top of itself (see below)
- Creating a felix86 related folder in /run/felix86

(Note that felix86 cannot use openat2+RESOLVE_IN_ROOT or rootless
namespaces for reasons, as much as we would like to)

Since there's a way to do fake mounts for trusted folders, we can use it
to fake mount /dev & co and no longer require on an actual `mount`, so
getting rid of that part was easy.

The bind mounting job of felix86-mounter was implemented for bubblewrap
-- bubblewrap will do `mount (NULL, "/", NULL, MS_SILENT | MS_SLAVE |
MS_REC, NULL)`. Which works normally, but if root (in our case, the
rootfs) isn't a mount it will fail. This is not a problem if the user
uses a rootfs that is actually a mount (for example if you are mounting
an ssd and that's your x86 rootfs) but if it's just a plain folder then
it will fail. A hack was implemented for now, where if this specific
mount fails it will do it to the host root instead. It should be fine to
do this in the case of bubblewrap since everything is namespaced, but it
is recommended the rootfs is an actual mount until a better solution
comes along. `sudo mount --bind $ROOTFS $ROOTFS` before running felix86
will work but is not required for bubblewrap to work.

The /run/felix86 folder was moved to /run/user/$UID/felix86 which
doesn't need higher privileges.

---

Once this is merged, felix86-mounter will no longer be needed and
felix86 will be able to run with no dependencies on root privileges.

---

More work needs to be done until this gets merged:
- [x] never relocate paths to be inside the rootfs (check if rootfs
substring or inside fakemount fds, else prompt for trust)
- [x] allow install script to just install a rootfs
- [x] add felix86 --shell option to replace `felix86 /bin/bash`
- [x] update documentation relating to felix86-mounter and felix86
/bin/bash
- [x] update install script to no longer install felix86-mounter