Pass config to children through environment variable and fix exit syscall from signal handler (#372)
Turns out child programs can have a different owner than what the
emulator was initially started with, and thus no permissions to open the
felix86 adjacent files.

For example, `sudo apt update` will run `apt-key`, and it won't be able
to access ~/.config/felix86/config.toml, which will make it exit before
it even starts.

This PR passes the config to children through an environment variable
instead. It also changes the permission of the named pipe to 0666 after
creating it.

---

There was also a bug with how exit/exit_group is handled when its ran
through a signal handler, and in some cases it would lead to stack
corruption. Calling exit/exit_group from the signal handler will now
correctly exit the process.

---

Fixes the apt-key issue in #371 
Fixes #373 
Fixes #365