Add support for execution profiles (#376)
Allows felix86 to run with different "execution profiles", which is
essentially a different configuration. This can be set with the
`FELIX86_PROFILE` environment variable. Any config values specified in
the profile will override the configurations in the config.toml.

This allows for quickly switching many options. For example, if a game
doesn't do anything out of the ordinary and follows the ABI,
`FELIX86_PROFILE=extreme` will enable all the optimizations. On the
other hand, `FELIX86_PROFILE=safe` will disable most optimizations
except the safe ones.

`FELIX86_PROFILE` either takes a name or an absolute path to a toml
file. If it's a name, it will search for that profile in
`$HOME/.config/felix86/profiles`. If it's a path, it will try to apply
the profile at that path if it exists.

---

To help with this, FELIX86_ENVIRONMENT is a new option. It can also be
specified from the config file/profiles. It allows passing environment
variables to the guest program separated by a ';'. An example usage is
`FELIX86_ENVIRONMENT="SDL_VIDEODRIVER=x11;DRI_PRIME=1"`.

---

This allows users to make custom profiles and pass the path to
`FELIX86_PROFILE`, which can enable or disable certain options in
felix86 and pass special environment variables.
`FELIX86_PROFILE=/path/to/profile.toml felix86 /path/to/my/game` for
example can enable optimizations that are safe with a specific game or
set environment variables that improve performance.